WIP Day 004
Cracking on, getting there, correct the mistakes that Casey makes.
Odd disconnect between using traditional gaming loops and Win32 messaging loops....to animate, you call RedrawWindow since you have the WM_PAINT already handling the drawing of the bitmap to screen, so all you need to do is post a WM_PAINT message to the screen. This means you need to invalidate the entire client rect of the current window, and then post a WM_PAINT.